+++ dashboards-comparator/dashboardscomparator-web/src/main/java/oracle/sysman/emaas/platform/dashboards/comparator/ws/rest/comparator/rows/entities/DashboardUserOptionsRowEntity.java
@@ -10,6 +10,8 @@
 
 package oracle.sysman.emaas.platform.dashboards.comparator.ws.rest.comparator.rows.entities;
 
+import java.math.BigInteger;
+
 import org.codehaus.jackson.annotate.JsonProperty;
 
 /**
@@ -24,7 +26,7 @@ public class DashboardUserOptionsRowEntity implements RowEntity
 	private Long tenantId;
 
 	@JsonProperty("DASHBOARD_ID")
-	private Long dashboardId;
+	private BigInteger dashboardId;
 
 	@JsonProperty("AUTO_REFRESH_INTERVAL")
 	private Long autoRefreshInterval;
@@ -162,7 +164,7 @@ public class DashboardUserOptionsRowEntity implements RowEntity
 	/**
 	 * @return the dashboardId
 	 */
-	public Long getDashboardId()
+	public BigInteger getDashboardId()
 	{
 		return dashboardId;
 	}
@@ -258,7 +260,7 @@ public class DashboardUserOptionsRowEntity implements RowEntity
 	 * @param dashboardId
 	 *            the dashboardId to set
 	 */
-	public void setDashboardId(Long dashboardId)
+	public void setDashboardId(BigInteger dashboardId)
 	{
 		this.dashboardId = dashboardId;
 	}
